split stuff out, make WindowDimensions its own thing
